+++ lib/database/src/so/elements/SoGLModelMatrixElement.cpp
@@ -257,7 +257,7 @@ SoGLModelMatrixElement::rotateEltBy(const SbRotation &rotation)
 
     rotation.getValue(axis, angle);
 
-    glRotatef(angle * (180.0f / M_PI), axis[0], axis[1], axis[2]);
+    glRotatef((GLfloat)(angle * (180.0 / M_PI)), axis[0], axis[1], axis[2]);
 }
 
 ////////////////////////////////////////////////////////////////////////
